Moves the generic encoding tests to the dedicated testbed project
