+++ testbeds/testbed-encoders/src/test/java/esa/mo/mal/encoder/binary/BaseBinaryTimeHandlerTest.java
